Refactor to use `Double` for `quantityCreativity` and `temperature` fields, add email sharing functionality, PDF generation for sharing messages, and enrich exception handling.

+++ src/main/java/com/kaust/ms/manager/prompt/chat/domain/models/responses/ChatResponse.java
@@ -20,7 +20,7 @@ public class ChatResponse {
     /** model. */
     private String model;
     /** quantityCreativity. */
-    private Integer quantityCreativity;
+    private Double quantityCreativity;
     /** createdAt. */
     private LocalDateTime createdAt;
     /** updatedAt. */
