+++ src/main/java/com/kaust/ms/manager/prompt/chat/infrastructure/mongodb/documents/HistoryActionsDocument.java
@@ -34,7 +34,7 @@ public class HistoryActionsDocument {
     /* model. */
     private String model;
     /* quantityCreativity. */
-    private Integer quantityCreativity;
+    private Double quantityCreativity;
     /* createdAt. */
     private LocalDateTime createdAt;
