+++ src/main/java/com/kaust/ms/manager/prompt/settings/infrastructure/mongodb/documents/SettingsDocument.java
@@ -21,7 +21,7 @@ public class SettingsDocument {
     /** model. */
     private Model model;
     /** quantityCreativity. */
-    private Integer quantityCreativity;
+    private Double quantityCreativity;
     /** createdAt. */
     private LocalDateTime createdAt;
     /** updatedAt. */
